+++ stars-core/src/main/kotlin/tools/aqua/stars/core/hooks/PreTSCEvaluationHook.kt
@@ -1,0 +1,45 @@
+/*
+ * Copyright 2024 The STARS Project Authors
+ * SPDX-License-Identifier: Apache-2.0
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package tools.aqua.stars.core.hooks
+
+import tools.aqua.stars.core.tsc.TSC
+import tools.aqua.stars.core.types.*
+
+/**
+ * A pre-evaluation hook that can be registered to a TSCEvaluation to be executed before the
+ * evaluation of the [TSC].
+ *
+ * @param E [EntityType].
+ * @param T [TickDataType].
+ * @param S [SegmentType].
+ * @param U [TickUnit].
+ * @param D [TickDifference].
+ * @param identifier The identifier to be used in the error message.
+ * @param evaluationFunction The function to be executed before the evaluation of the [TSC].
+ */
+open class PreTSCEvaluationHook<
+    E : EntityType<E, T, S, U, D>,
+    T : TickDataType<E, T, S, U, D>,
+    S : SegmentType<E, T, S, U, D>,
+    U : TickUnit<U, D>,
+    D : TickDifference<D>>(
+    identifier: String,
+    evaluationFunction: (TSC<E, T, S, U, D>) -> EvaluationHookResult
+) :
+    EvaluationHook<TSC<E, T, S, U, D>>(
+        identifier = identifier, evaluationFunction = evaluationFunction)