+++ stars-core/src/main/kotlin/tools/aqua/stars/core/metric/metrics/postEvaluation/FailedMonitorsGroupedByTSCInstanceMetric.kt
@@ -18,13 +18,15 @@
 package tools.aqua.stars.core.metric.metrics.postEvaluation
 
 import java.util.logging.Logger
-import tools.aqua.stars.core.metric.metrics.evaluation.ValidTSCInstancesPerProjectionMetric
+import tools.aqua.stars.core.metric.metrics.evaluation.ValidTSCInstancesPerTSCMetric
 import tools.aqua.stars.core.metric.providers.Loggable
 import tools.aqua.stars.core.metric.providers.PostEvaluationMetricProvider
+import tools.aqua.stars.core.metric.utils.ApplicationConstantsHolder.CONSOLE_INDENT
+import tools.aqua.stars.core.metric.utils.ApplicationConstantsHolder.CONSOLE_SEPARATOR
+import tools.aqua.stars.core.tsc.TSC
 import tools.aqua.stars.core.tsc.TSCFailedMonitorInstance
 import tools.aqua.stars.core.tsc.instance.*
 import tools.aqua.stars.core.tsc.node.TSCNode
-import tools.aqua.stars.core.tsc.projection.TSCProjection
 import tools.aqua.stars.core.types.*
 
 /**
@@ -32,15 +34,15 @@ import tools.aqua.stars.core.types.*
  * [TSCNode.monitorFunction]s that evaluate to 'false'.
  *
  * This class implements the [Loggable] interface. It logs and prints the count and names of all
- * failing [TSCNode.monitorFunction]s for each [TSCProjection]. It logs the failing
- * [TSCFailedMonitorInstance]s for each [TSCProjection] and groups it by the [TSCInstance].
+ * failing [TSCNode.monitorFunction]s for each [TSC]. It logs the failing
+ * [TSCFailedMonitorInstance]s for each [TSC] and groups it by the [TSCInstance].
  *
  * @param E [EntityType].
  * @param T [TickDataType].
  * @param S [SegmentType].
  * @param U [TickUnit].
  * @param D [TickDifference].
- * @property dependsOn The instance of a [ValidTSCInstancesPerProjectionMetric] on which this metric
+ * @property dependsOn The instance of a [ValidTSCInstancesPerTSCMetric] on which this metric
  *   depends on and needs for its calculation.
  * @property logger [Logger] instance.
  */
@@ -51,24 +53,24 @@ class FailedMonitorsGroupedByTSCInstanceMetric<
     S : SegmentType<E, T, S, U, D>,
     U : TickUnit<U, D>,
     D : TickDifference<D>>(
-    override val dependsOn: ValidTSCInstancesPerProjectionMetric<E, T, S, U, D>,
+    override val dependsOn: ValidTSCInstancesPerTSCMetric<E, T, S, U, D>,
     override val logger: Logger = Loggable.getLogger("failed-monitors-grouped-by-tsc-instance")
 ) : PostEvaluationMetricProvider<E, T, S, U, D>, Loggable {
 
   /**
-   * Holds a [Map] from a [TSCProjection] to a [Map] from a node label (as [String], representing
-   * the ID of the monitor) to a [Map] from a [TSCInstanceNode] to a [List] of all occurring
+   * Holds a [Map] from a [TSC] to a [Map] from a node label (as [String], representing the ID of
+   * the monitor) to a [Map] from a [TSCInstanceNode] to a [List] of all occurring
    * [TSCInstanceNode]s.
    */
   private val failedMonitors:
       MutableMap<
-          TSCProjection<E, T, S, U, D>,
+          TSC<E, T, S, U, D>,
           Map<String, Map<TSCInstanceNode<E, T, S, U, D>, List<TSCInstance<E, T, S, U, D>>>>> =
       mutableMapOf()
 
   /**
    * Calculates a grouped [Map] of [TSCFailedMonitorInstance]s with grouped [TSCInstance]s for all
-   * [TSCProjection]s by validating all monitors for all valid [TSCInstance]s.
+   * [TSC]s by validating all monitors for all valid [TSCInstance]s.
    */
   override fun postEvaluate() {
     failedMonitors.clear()
@@ -87,12 +89,14 @@ class FailedMonitorsGroupedByTSCInstanceMetric<
         })
   }
 
-  /** Prints the count of failed monitors for each [TSCProjection]. */
+  /** Prints the count of failed monitors for each [TSC]. */
   override fun printPostEvaluationResult() {
-    failedMonitors.forEach { (projection, failedMonitors) ->
+    println(
+        "\n$CONSOLE_SEPARATOR\n$CONSOLE_INDENT Failed Monitors Grouped By TSC \n$CONSOLE_SEPARATOR")
+    failedMonitors.forEach { (tsc, failedMonitors) ->
       if (failedMonitors.isEmpty()) return@forEach
 
-      logInfo("Failed monitors for projection '$projection':")
+      logInfo("Failed monitors for tsc '${tsc.identifier}':")
 
       failedMonitors.forEach { monitor ->
         logInfo(