+++ stars-core/src/main/kotlin/tools/aqua/stars/core/metric/providers/TSCMetricProvider.kt
@@ -18,14 +18,12 @@
 package tools.aqua.stars.core.metric.providers
 
 import tools.aqua.stars.core.evaluation.TSCEvaluation
-import tools.aqua.stars.core.tsc.instance.TSCInstance
-import tools.aqua.stars.core.tsc.projection.TSCProjection
+import tools.aqua.stars.core.tsc.TSC
 import tools.aqua.stars.core.types.*
 
 /**
- * The [ProjectionAndTSCInstanceNodeMetricProvider] implements the [EvaluationMetricProvider] and
- * provides an [evaluate] function which gets a [TSCProjection] and a [TSCInstance] which is called
- * during the evaluation phase.
+ * The [TSCMetricProvider] implements the [EvaluationMetricProvider] and provides an [evaluate]
+ * function which gets a [TSC] which is called during the evaluation phase.
  *
  * @param E [EntityType].
  * @param T [TickDataType].
@@ -34,7 +32,7 @@ import tools.aqua.stars.core.types.*
  * @param D [TickDifference].
  * @see TSCEvaluation.runEvaluation
  */
-interface ProjectionAndTSCInstanceNodeMetricProvider<
+interface TSCMetricProvider<
     E : EntityType<E, T, S, U, D>,
     T : TickDataType<E, T, S, U, D>,
     S : SegmentType<E, T, S, U, D>,
@@ -42,10 +40,10 @@ interface ProjectionAndTSCInstanceNodeMetricProvider<
     D : TickDifference<D>> : EvaluationMetricProvider<E, T, S, U, D> {
 
   /**
-   * Evaluate the metric based on the given parameters.
+   * Evaluate the metric based on the given parameter.
    *
-   * @param projection The current [TSCProjection].
-   * @param tscInstance The current [TSCInstance].
+   * @param tsc The current [TSC].
+   * @return The evaluation result.
    */
-  fun evaluate(projection: TSCProjection<E, T, S, U, D>, tscInstance: TSCInstance<E, T, S, U, D>)
+  fun evaluate(tsc: TSC<E, T, S, U, D>): Any?
 }